requirements.txt
telethon==1.24.0
gspread==5.7.2
oauth2client==4.1.3
python-dotenv==1.0.0
